# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus Chat Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed

  # ─── USERS ──────────────────────────────────────────────────────────
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, banned]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Suspended
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Banned
  /admin/users/{userId}/activate:
    post:
      operationId: activateUser
      tags: [admin]
      summary: Activate a user account
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Activated
  /admin/reports:
    get:
      operationId: listReports
      tags: [admin]
      summary: List user and content reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved, dismissed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /admin/broadcast:
    post:
      operationId: adminBroadcast
      tags: [admin]
      summary: Send platform-wide announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "204":
          description: Sent
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get platform settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:    { type: string, minLength: 2 }
        email:       { type: string, format: email }
        password:    { type: string, minLength: 8 }
        displayName: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword:     { type: string, minLength: 8 }

    # ─── USERS ───────────────────────────────────────────────────────
    User:
      type: object
      required: [id, username, email, displayName, status, role, createdAt]
      properties:
        id:          { type: integer }
        username:    { type: string }
        email:       { type: string }
        displayName: { type: string }
        avatarUrl:   { type: ["string", "null"] }
        bio:         { type: ["string", "null"] }
        status:      { type: string, enum: [active, suspended, banned] }
        presence:    { type: string, enum: [online, away, offline] }
        role:        { type: string, enum: [user, moderator, admin] }
        isVerified:  { type: boolean }
        lastSeenAt:  { type: ["string", "null"] }
        createdAt:   { type: string }
    UserUpdate:
      type: object
      properties:
        displayName: { type: string }
        bio:         { type: string }
        avatarUrl:   { type: string }
        presence:    { type: string, enum: [online, away, offline] }

    # ─── CONVERSATIONS ───────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, username, email, displayName, status, role, createdAt]
      properties:
        id:           { type: integer }
        username:     { type: string }
        email:        { type: string }
        displayName:  { type: string }
        avatarUrl:    { type: ["string", "null"] }
        status:       { type: string, enum: [active, suspended, banned] }
        role:         { type: string, enum: [user, moderator, admin] }
        messageCount: { type: integer }
        lastSeenAt:   { type: ["string", "null"] }
        createdAt:    { type: string }
    Report:
      type: object
      required: [id, reporterId, reason, status, createdAt]
      properties:
        id:           { type: integer }
        reporterId:   { type: integer }
        reporter:
          $ref: "#/components/schemas/User"
        targetUserId: { type: ["integer", "null"] }
        targetUser:
          $ref: "#/components/schemas/User"
        messageId:    { type: ["integer", "null"] }
        reason:       { type: string }
        description:  { type: ["string", "null"] }
        status:       { type: string, enum: [pending, resolved, dismissed] }
        createdAt:    { type: string }
    AnnouncementInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string }
        body:  { type: string }
    PlatformSettings:
      type: object
      required: [emailRegistration, maintenanceMode]
      properties:
        emailRegistration:  { type: boolean }
        maintenanceMode:    { type: boolean }
        maxFileUploadMb:    { type: integer }
        allowGuestBrowse:   { type: boolean }
        requireEmailVerify: { type: boolean }
    PlatformSettingsUpdate:
      type: object
      properties:
        emailRegistration:  { type: boolean }
        maintenanceMode:    { type: boolean }
        maxFileUploadMb:    { type: integer }
        allowGuestBrowse:   { type: boolean }
        requireEmailVerify: { type: boolean }
...