# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPresence } from './userPresence';
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  email: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  status: UserStatus;
  presence?: UserPresence;
  role: UserRole;
  isVerified?: boolean;
  /** @nullable */
  lastSeenAt?: string | null;
  createdAt: string;
}