# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdatePresence } from './userUpdatePresence';

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  presence?: UserUpdatePresence;
}